V1/core/services/whatsapp-service/requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
qrcode==7.4.2
pillow==10.1.0
